feat: refactoring CardSizes
