+++ src/entities/Card/index.ts
@@ -1,4 +1,6 @@
+export { CardColorChanger } from './CardColorChanger/CardColorChanger';
 export { CardDescription } from './CardDescription/CardDescription';
 export { CardImage } from './CardImage/CardImage';
+export { CardSizesChanger } from './CardSizesChanger/CardSizesChanger';
 export { FullCard } from './FullCard/FullCard';
 export { ShortCard } from './ShortCard/ShortCard';
